requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
llama-cpp-python[cuda]>=0.2.73
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv
starlette
py-cpuinfo
langtrace-python-sdk>=2.0.0
